+++ server/src/routes/doctorRoutes.ts
@@ -3,10 +3,8 @@ import DoctorController from "../controllers/doctorController.js";
 
 const router: Router = Router(); 
 
-router.get('/get/:id', DoctorController.getOne);
-
-router.get('/get', DoctorController.getAll);
-
+router.get('/:id', DoctorController.getOne);
+router.get('/:userId', DoctorController.getDoctorByUser);
 router.post('/create', DoctorController.create);
 
 export default router;
