+++ server/src/routes/userRoutes.ts
@@ -16,5 +16,6 @@ router.get('/activate/:link', UserController.activate);
 router.get('/refresh', UserController.refresh);
 router.get('/:id', UserController.getUser);
 router.post('/check', UserController.checkUser);
+
 export default router;
